Coding Days: Fixed conversion warnings in parallel
git-svn-id: https://ipvs.informatik.uni-stuttgart.de/SGpp/repos/trunk@2658 4eea3252-f0fb-4393-894d-40516dce545b

+++ src/sgpp/parallel/datadriven/algorithm/DMWeightMatrixSPVectorizedIdentity.cpp
@@ -14,7 +14,7 @@ namespace sg
 namespace parallel
 {
 
-DMWeightMatrixSPVectorizedIdentity::DMWeightMatrixSPVectorizedIdentity(sg::base::Grid& SparseGrid, sg::base::DataMatrixSP& trainData, double lambda, sg::base::DataVectorSP& w, VectorizationType vecMode)
+DMWeightMatrixSPVectorizedIdentity::DMWeightMatrixSPVectorizedIdentity(sg::base::Grid& SparseGrid, sg::base::DataMatrixSP& trainData, float lambda, sg::base::DataVectorSP& w, VectorizationType vecMode)
 {
 	// handle unsupported vector extensions
 	// @TODO (heinecke) find a better way to determine vectorwidth
@@ -118,7 +118,7 @@ void DMWeightMatrixSPVectorizedIdentity::generateb(sg::base::DataVectorSP& class
 	//resize the data class vector and weight
 	if (numPatchedTrainingInstances != numTrainingInstances)
 	{
-		double lastClass;
+		float lastClass;
 		for (size_t i = 0; i < loopcount; i++)
 		{
 			lastClass = myClassesWithWeights.get(myClassesWithWeights.getSize()-1);
